# api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
...
paths:
...
  /api/v1/auth/signup:
    post:
      tags:
      - Registration and Authentication controller
      summary: SignUp
      description: Signup new user
      operationId: signUp
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignUpRequest"
        required: true
      responses:
        "201":
          description: "Object, containing new user and access token"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SignUpResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppErrorResponse"
              example:
                status: 400
                message: Validation Error
                timestamp: 2024-06-04T22:40:46.7924577
  /api/v1/auth/signin:
    post:
      tags:
      - Registration and Authentication controller
      summary: SignIn
      description: Signin existed user
      operationId: signIn
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignInRequest"
        required: true
      responses:
        "200":
          description: "Object, containing access token"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SignInResponse"
              example:
                accessToken: eyJhbGciOiJIUzM4NCJ9.eyJpYXQiOjE3MTc1Mjc1NTUsImV4cCI6MTcxNzUyODE1NSwic3ViIjoiTWFsdmluYSIsInJvbGVzIjpbIlJPTEVfVVNFUiJdfQ.utQS9pSI-CBEx-s8P2O1YsVwe7ofjQLX-YAj8b3yZ9Y-817TXbzlnuUwdOqahKHX
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppErrorResponse"
              example:
                status: 400
                message: Validation Error
                timestamp: 2024-06-04T22:40:46.7924577
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppErrorResponse"
              example:
                status: 401
                message: "Authentication error: 'Bad credentials'"
                timestamp: 2024-06-04T22:40:46.7924577
...
components:
  schemas:
...
    UserResponseDtoFull:
      type: object
      properties:
        id:
          type: integer
          description: User Id
          format: int64
          example: 1
        email:
          type: string
          description: User email
          example: user@domain.com
        username:
          type: string
          description: Username
          example: User1
      description: User response DTO with full info
    AppErrorResponse:
      type: object
      properties:
        status:
          type: integer
          description: HTTP Status
          format: int32
          example: 401
        message:
          type: string
          description: Message
          example: Access token expired
        timestamp:
          type: string
          description: Timestamp
          format: date-time
      description: App error response
...
    SignUpResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/UserResponseDtoFull"
        accessToken:
          type: string
          description: Access token
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.Sf...
      description: SignUp response with new User and Access token
    SignUpRequest:
      required:
      - email
      - password
      - username
      type: object
      properties:
        email:
          maxLength: 50
          minLength: 6
          type: string
          description: User email
          example: user@domain.com
        password:
          maxLength: 20
          minLength: 1
          type: string
          description: Password
          example: mY_password
        username:
          maxLength: 50
          minLength: 3
          type: string
          description: Username
          example: User1
      description: SignUp request
    SignInResponse:
      type: object
      properties:
        accessToken:
          type: string
          description: Access token
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.Sf...
      description: SignIn response with Access token
    SignInRequest:
      required:
      - email
      - password
      type: object
      properties:
        email:
          maxLength: 50
          minLength: 6
          type: string
          description: User email
          example: user@domain.com
        password:
          maxLength: 20
          minLength: 1
          type: string
          description: Password
          example: mY_password
      description: SignIn request
...